Add data input from JSON functions
Adds functions that takes input data from Scancode Scan Results, in JSON, and
structures them similarly as the ClearlyDefined DataBase Data, so the same
fuctions can be used on them. Adds Jupyter Notebook to explain the Fuction Calls,
and Data, and JSON files as sample, from the issues in nexB/scancode-toolkit#1963.

Signed-off-by: Ayan Sinha Mahapatra <ayansmahapatra@gmail.com>